revised extension interface
Extensions conceptionally don't belong to the kernel but to the array.
This becomes clear when trying to figure out multi-window kernels that
may have different extensions per argument array.

This commit introduces the `extend(::AbstractArray, ::Extension)` method
to create an array wrapper for specifying the extension and removes the
previous kernel-based construction, thus paving the way for multi-window
kernel application.